Implement cassandra-to-cassandra, cassandra-to-elasticsearch and cassandra-to-kafka routes

+++ src/main/scala/io/oss/data/highway/utils/FilesUtils.scala
@@ -159,9 +159,12 @@ object FilesUtils {
     * @return Array of Unit
     */
   def cleanup(in: String): Array[Unit] = {
-    new File(in)
-      .listFiles()
-      .map(FileUtils.forceDelete)
+    if (new File(in).exists())
+      new File(in)
+        .listFiles()
+        .map(FileUtils.forceDelete)
+    else
+      Array()
   }
 
   /**
